backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
pymysql
python-jose[cryptography]
